+++ detail/client/src/main/java/au/com/deanpike/detail/client/usecase/PropertyDetailUseCase.kt
@@ -4,5 +4,5 @@ import au.com.deanpike.commonshared.util.ResponseWrapper
 import au.com.deanpike.detail.client.model.detail.PropertyDetail
 
 interface PropertyDetailUseCase {
-    suspend fun getPropertyDetails(id: Int): ResponseWrapper<PropertyDetail>
+    suspend fun getPropertyDetails(id: Long): ResponseWrapper<PropertyDetail>
 }
