+++ detail/data/src/main/java/au/com/deanpike/detail/data/usecase/PropertyDetailUseCase.kt
@@ -9,7 +9,7 @@ import javax.inject.Inject
 internal class PropertyDetailUseCaseImpl @Inject constructor(
     private val repo: PropertyDetailRepository
 ) : PropertyDetailUseCase {
-    override suspend fun getPropertyDetails(id: Int): ResponseWrapper<PropertyDetail> {
+    override suspend fun getPropertyDetails(id: Long): ResponseWrapper<PropertyDetail> {
         return repo.getDetails(id)
     }
 }
